Add v2 currency utils
