+++ src/hooks/v2/V2ContractLoader.ts
@@ -36,7 +36,7 @@ export function useV2ContractLoader() {
 
         setContracts(newContracts)
       } catch (e) {
-        console.log('CONTRACT LOADER ERROR:', e)
+        console.error('CONTRACT LOADER ERROR:', e)
       }
     }
